+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ### 0.2.1
 
+* `bijection-guava` for Guava interop.
+
 ### 0.2.0
 
 * `bijection-scrooge` for scrooge generated scala thrift code